requirements.txt
asttokens==2.0.5
backcall==0.2.0
click==8.1.3
decorator==5.1.1
executing==0.8.3
Flask==2.1.3
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.6.1
parso==0.8.3
pdf2image==1.16.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
Pygments==2.12.0
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
stack-data==0.3.0
traitlets==5.3.0
wcwidth==0.2.5
Werkzeug==2.1.2
zipp==3.8.1
